doc: add documentation for the library
The documentation is to be rendered (primarily as HTML) with Sphinx. A
specific Makefile is available in the doc/ directory, offering targets
like "make html". It can also be called via the top-level Makefile, with
"make doc".

This documentation is expected to be published on Read The Doc.

Thanks to @pcasconnetronome for spotting many mistakes in the
documentation, now fixed.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>