Makefile: create a Makefile to build the project
Add a Makefile to build and install the library.

Pass a version number to the library when building it.

Library is built with symbols hidden by default.

Some environment variables can be used to silence the build ($Q) or to
disable the numerous extra warnings used in the Makefile ($W).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>