lib: add support for rules with multiple match patterns
So far we had support for different rules, so one could for example try
to block:

* packets coming from IP foo (one rule),
* or packets going to port bar (another rule).

It was not possible to block packets coming from IP foo AND going to
port bar at the same time (combination of patterns).

This commit adds support for rules with several match patterns.

We cannot, however, have an unlimited number of patterns, because we do
not have the possibility to loop over a list of patterns in the BPF
program. Therefore, we fix a limit of 5 patterns for now. This is a
maximum value, and does not mean that all rules (or any rule in the
filter, for that matter) have to contain that many patterns. The C code
generation tries to minimize the amount of code generated, depending on
the maximum number of patterns found for the rules in a filter.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>