proggen: annotate the BPF map for BTF support
Makes bpftool able to dump the structure of the map, if the BPF program
is compiled with -g flag (libkefir does that) and with a recent LLVM
version (8+).

For the recored, under the same conditions (but without the need for any
additional annotation), the C instructions should be passed as BTF
information to the kernel.

All this debug information can be inspected with e.g. bpftool dumps.

Note that as of this writing, BTF is not supported with programs
offloaded to the hardware, and trying to load BTF info for the maps will
display a warning (then libbpf runs a second attempt without BTF, which
is supposed to work).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>