parse_ethtool: rework parsing of integers, move definitions
Change some of the formats used by the ethtool rules parsing facility to
make it more flexible. Update the filter dump facility accordingly.

Additionally, move the integer parsing functions to a new
libkefir_parse.c file, so that they can be reused by other parsing
module in the future.

The error functions are similarly moved to the header related to this
new file, as a macro that can take a component name as an argument and
defines the two logging helpers from that (for errors and for bugs).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>