parse_ethtool: change rule structure to track match_type
The header_type / match_offset / match_length tuple is somewhat
flexible, but at the end of the day we loose too much information for
some kinds of matches (e.g. what about non-fixed offset?), so just keep
track of the very field the match is based on. Update the dump functions
and ethtool parsing accordingly (and simplify the parsing for ethtool,
which suffered from a bit of overthinking).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>