proggen: generate more BPF code, to be able to apply filters
Extend the code generation to produce a functional BPF filtering
program. Key extraction, including IPv4/IPv6/L4 headers processing, can
now be used to retrieve metadata on a packet. These keys are then used
compared again each rule of the filter to see if the packet matches one
of them, and in that case, the relevant action is performed.

Support of all rules is not complete, but the main items (IP addresses,
L4 ports, a few additional ones) work.

Move (and rename with "KEFIR_" as prefix) the definitions related to
cprog options to libkefir_internals, as options will be needed for
loading and attaching the program as well.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>