libkefir: add kefir_cprog_to_buf() to dump C code
The function takes a C program object and dumps the relevant C code for
the BPF program to a buffer. The user can use for instance printf() to
dump this buffer into the console.

If the buffer provided by the user is too small to embed the whole
program, it is realloc()'ed.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>