compile: exec() to clang/llc to compile the C code into BPF
Add a function to the library so that the user can compile a C file
(previously generated with a call to kefir_cprog_to_file()) into an ELF
object file containing the BPF bytecode related to the initial filter.

Right now, the implementation relies on fork()s and exec()s into the
clang and llc programs directly, so those programs must be available on
the system for the function to succeed.

The function allows for custom paths to the clang and llc executables
(if NULL values are passed, a default path is used).

If the name of the object file (or the intermediary file containing LLVM
intermediate representation) to produce is NULL, then it is derived from
the input C source file (assuming the latter has a ".c" extension).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>