proggen: rename two functions (prefix the names with "proggen")
Makes naming more consistent.

As a rule of thumbs, all functions exposed to the user are defined in
libkefir.c (and declared in libkefir.h) and start with "kefir_".
Additionally, as is of concern here, all functions exposed from a file
to the main libkefir.c library file start with a prefix derived from the
name of that file (no need to prefix them with "kefir_", they are not
API).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>