compile: link to libbpf, add functions to load and attach the program
Add functions relying on libbpf to load or load-and-attach a BPF program
previously generated from a filter object and compiled as an object file
(both the internal cprog object and a path to the generated object file
are required).

On attaching the program, the rules related to the filter are loaded
into the BPF map used to that purpose by the program.

For now, only XDP is supported for attaching programs.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>