parse: parse masks for TC flower and ethtool rules
Implement parsing for masks provided as parts of TC flower or ethtool
rules. Masks are stored in the filter object, but are not yet used in
the generated C/eBPF program at this point.

Also fix prototype of parse_ipv6_addr(), which takes a pointer to a
uint32_t instead of a uint8_t.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>