proggen: use masks in generated C program
Support for parsing masks was recently added to the library. Masks are
stored in the filter object. But so far they were not used in the
generated program.

Add masks to the rules passed to the BPF program through the rules
table, and apply them to the values to check for deciding what to do of
the packet.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>