compile, proggen: make masks conditional in generated program
Using masks and applying them to each value to compare in the BPF
program has a huge cost in terms of number of instructions (hence in
performance, and in terms of maximum size for the program, especially
for offload).

To reduce the impact, do not use masks (at all) in BPF programs when no
rule in the filter needs any mask. This makes the generated program
cleaner when no mask is used, at the expense of the generating code in
the library, which gets more complex.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>