json_save: save a filter object to a file as JSON
Implement an API function used to save a filter object to an external
file, formatted as JSON. Next step will be to implement a function to
reload this saved filter.

A version number for the library is added to the JSON object, to help
with compatibility issues when reloading the file in the future.
However, libkefir does not have any version numbering mechanism yet, so
this is not complete (defaults to a "WIP" value).

JSON formatting is provided by the json_writer mini-library, extracted
from iproute2.

File json_writer.h received one change: it now includes <stdio.h>
(necessary because of the use of type FILE *).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>