compile: reset errno after pause()
When the library resumes after pausing to wait for the forked threads to
finish executing clang or llc, the errno variable is set to EINTR
(interrupt). This is expected, so it brings no useful information. On
the contrary, if a process calling libkefir functions checks errno and
assumes that it should be at 0 to indicate that everything worked fine,
the EINTR error may look confusing.

Simply reset errno to 0 after resuming from pause(), to avoid bad
surprises.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>