internals: share ARRAY_SIZE, offsetof, sizeof_member
The ARRAY_SIZE, offsetof, and sizeof_member macros, even if used only in
ethtool parsing so far, are not specially related to ethtool parsing.
We will soon reuse one of them when saving/restoring a filter object to
and from a file.

Move these macros to the common header file in order to makes them
visible to all components.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>