libkefir: implement kefir_delete_rule_by_id() function
Implement the API function allowing to delete a rule from a filter
object, at a given location in the list.

Note that the id must be the index of the rule in the list as stored in
the filter. In case we process the list in the future and change the
order after the user loads a rule, this may NOT be the order in which
the user loaded the rule (e.g. if we removed duplicates in the list).

Therefore, we will need a way in the future to make sure the user is
aware of the id of the list. We could make the loading function
returning the id of the whent it is loaded, but then it would be subject
to change if additional rules are loaded after that. We could dump all
the filter and print ids associated to the rules, but that will not be
really helpful in terms of programmability. One solution could be to add
an attribute to each rule to keep the order in which it was added, so we
could instead match on this user id in the future.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>