compile: make libbpf print its messages to libkefir error buffer
This is an experiment, and may or may not prove to be a good idea: We
prevent libbpf from printing its own messages to the console. Instead,
we inject a custom printing function that append all messages to
libkefir error string. This way, if the user prefers to handle this
string in a specific way, they will also have access to libbpf output.
And at the same time, this prevents libbpf from "polluting" console
ouput if no output is expected at all.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>