compile: make load and attach forward compatible (use attr argument)
The list of arguments needed to load and attach programs includes flags
or interface index under some circumstances. It could include other
attributes in the future, depending on how BPF evolves in the Linux
kernel.

To improve forward compatibility, remove ifindex and flags from the
argument list of related functions, in particular the ones exposed in
the API. Instead, pass a pointer to a dedicated struct that encompasses
all of those attributes.

Since we changed it, we can now easily add a log_level attribute to the
list, to benefit from the latest changes in libbpf and dump verifier
debug output if the user wishes to do so.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>