json_restore: load a filter saved as JSON from a file
The library previously received support for saving a filter object as
JSON into an external file. This commit implements the revert process,
and generates a new filter object fields with the rules loaded from that
file.

The file is supposed to have been generated from a previous call to the
library. It is NOT supposed to be manually created by users. Users could
create or edit a file at their own risk; the library does not makes
attempts at being very flexible in parsing, so the structure of the JSON
file would have to match exactly what the restoration function expects.

A few attempts are made, however, to possibly ease forward
compatibility.

As noted before about saving the program, the version numbering
mechanism for the library is yet to be implemented. The current function
does expect and parse a version string in the file containing the
filter, but nothing is done with that string at this time.

JSON parsing is provided by the JSMN mini-library.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>