internals: fix type for kefir_cprog_options
A new type was previously defined for struct kefir_cprog_options, and
used in struct kefir_cprog. Since this type was not use anywhere else,
and in particular not used by users of the lib, it was removed in a
recent commit. Fix struct kefir_cprog accordingly: use the complete
"struct kefir_cprog_options" type.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>