libkefir: always pass non-zero ifindex for attaching the program
A bug was introduced in a previous commit, when the API for loading and
attaching programs changed: the ifindex, that should be at 0 for loading
programs when no hardware offload is required, was similarly set to 0
for attaching the program as well when it is not offloaded.

This cannot work, since we do need a non-zero ifindex for attaching the
program. Fix this by backing up the value of the ifindex (as provided by
the user) before setting it to zero, when there is no offload, in attr,
and setting it back before trying to attach the program.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>