compile: reset errno after calling bpf_prog_load_xattr()
Libbpf performs a number of operations for loading the program, some of
which may set errno to an error value even if all the processing went as
expected. This may be confusing for users observing the value of errno.

If program loading went just fine, i.e. if the function returns zero,
unset errno.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>