json_restore: reset errno before calling strtol()
The success of strtol() is checked by observing the value of errno after
the call. However, strtol() does not reset errno on success, it simply
leaves it unchanged. This means that if errno was set prior to the call,
then the validation check after strtol() will fail, and we will exit
early from JSON parsing.

Fix it by resetting errno just before calling strtol().

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>