libkefir: return early from kefir_load_rule() if rule creation failed
If creating a rule failed for some reason, typically because parsing the
rule failed, return early from kefir_load_rule() and do not attempt to
add the NULL rule to the rule set.

This fixes a segmentation error that would occur when trying to
dereference a rule that was not created correctly.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>