error: fix max size of message to append to kefir error string
In error_vappend_str(), we append a message to the error string of the
library. In order to prevent buffer overflow we use a maximum size for
the number of bytes to write, however this limit is not correct.

Fix it by subtracting the offset at which we write (plus one byte for
the terminating nul byte).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>