parse_ethtool, parse_tc: initialise memory when allocating rules
Rule allocation did not require initialisation when it was first
implemented, because rules only had one match object, which would be
written in all cases.

However, now that we support several match objects per rule, and that
some of those objects can be left at zero, rules require to be set at
zero initially. To that end, replace the calls to malloc() with
calloc().

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>