parse, proggen: reorganise L4 ports match handling
So far, the library would use a single match type for L4 source port (or
another one for destination port), regardless of the requirements on IP
version. After all, we can have rules focusing only on L4 ports, and not
taking into account if the packet is IPv4 or IPv6.

However, this is in contradiction with the usual behaviour of tc and
ethtool rules. Both rule formats takes a flow type (ethtool) or protocol
(tc) to indicate to what types of packets (typically IPv4 VS. IPv6) the
rule with port matching should apply.

So we need a way to differenciate IP version when matching port. This
could be implemented as a second match on ethertype, but that would be
rather cumbersome (and adding matches can raise the max number of
matches in the filter, making the BPF programs longer and less
efficient), so it's better if we can avoid that.

Instead, dissociate the different cases by adding several match types:
one for e.g. source port for IPv4 packets, one for IPv6 packets, and one
for any IP version. In this way, we do not have to juggle with a second
match to generate the relevant code.

On the BPF front, add conditions on the packet being IPv4 or IPv6 (we
already parsed ethertype anyway) when checking ports for IPv4 only or
IPv6 only packets.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>