parse: fix "/24"-style masks parsing
Calling parse_uint() for masks would store them in the specific fashion
used for storing values (in a way optimized for later comparisons in the
BPF programs), but would actually result in passing an erroneous value
to the function supposed to turn it into a bitmask.

Fix it by re-doing the strtoul() processing in parse_slash_mask().

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>