proggen: fix length for eth_proto and nh_off variables
eth_proto (ethertype) should be two byte long, but was only a uint8_t!
Let's fix it.

Similarly, nh_off (next header offset) is four byte long, which is not
an error, but not something necessary either---we do not need more than
one byte to hold the value. Let's change it as well.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>