error: provide a function to reset the error string of the library
The library is using a static buffer to hold error messages. In order to
gather all error messages in that single buffer, all messages from
libbpf are incrementally appended to this buffer.

This may have an undesirable effect: When libbpf prints warnings that we
ignore (Typically: because loading a BTF object failed), the error
string grows and grows, and we get a long succession of warnings after
a while.

To mitigate this, provide a function that can be used to reset the error
string to an empty string (and reset the internal index, so we start
appending from the beginning of the buffer again).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>