parse_ethtool, proggen: complete VLAN support
Complete VLAN parsing in ethtool (tag, ethertype) and add the relevant
code to C program generation.

Note that support is not perfect yet: ethtool rules can use vlan as an
"extended" attribute, in the sense it can be added to another match
object. For now, the library only supports a VLAN match as the only
match element for an ethtool rule.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>