parse_ethtool: accept several matches to account for L4 flow types
The rules for ethtool specify a flow-type that, in addition to indicate
if the match should be done on IPv4 or IPv6, can indicate a given L4
protocol (UDP, TCP, or SCTP).

Change the way rules are built in ethtool parsing to account for these
flow-types, in a manner that adds another match object for L4 protocol
whenever this is relevant.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>