proggen: define and use a macro to generate/append code to buffer
Define a macro that appends text to a buffer, and use it everywhere in
the file to add generated code to the C program in construction.

This macro does NOT follow what we could consider good practice:

- It "hides" a conditional "return -1;", if appending fails. This makes
  it less obvious that we can return early from the function using the
  macro.
- It uses local variable names, "buf" and "buf_len", that have to be
  defined and to have the correct semantic in the context where the
  macro is used.

Due to the high number of calls to buf_append() that could be replaced
by this macro, the macro is introduced with the hope that the gain in
terms of readability (indentation, visual flow with much fewer
structures "if (...) return -1;" will overweight the drawbacks mentioned
above.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>