parse_ethtool: add support for masks
Support passing masks in ethtool rules. We already had program
generation from rules with mask, and parsing of masks for TC rules, but
ethtool support was missing: let's add it.

We reuse (but rename) the function used for parsing values. Make it take
a void * for the destination where to store the value or mask, so we can
pass a pointer to either a value object or a mask object.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>