parse_ethtool: remove mask at value parsing
This mask was intended to be used for collecting the value from the
packet at the correct offset. It is in fact unnecessary, because the BPF
program generated can handle collecting without the need for such masks.

Worse, if no other mask is added, this is a leftover mask that 1) will
trigger mask usage for all values in the program, even if there is no
other use of masks for the current filter, and 2) may render check
result incorrect.

Let's simply remove it.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>