error: remove "parsing" from component names generation
A leftover from the time when the DEFINE_ERR_FUNCTIONS() macro was used
for parsing components only, the "parsing" word was part of all
generated component names (for error strings).

This is not desirable. Move that word to the components for which it is
relevant, ethtool and tc rules parsing, and remove it from the macro so
that it does not appear for non-parsing components.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>