parse: parse a bitmask, not a prefix length in parse_uint_slash_mask()
Function parse_uint_slash_mask() is used to parse an integer value
possibly followed by a slash and a mask. This mask should be a bitmask
(e.g. "/0xff"), not a prefix length (e.g. "/24" for IPv4 addresses).
Therefore, the mask value should be parsed as a regular int, and does
not require additional convertion from prefix length to bitmask. Fix the
function accordingly.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>