parse: fix parsing of prefix length masks for ethernet addresses
The check for successful detection of prefix length mask for ethernet
addresses is incorrect: parse_uint() returns 0 on success, and in that
case only we want to proceed with the value parsed for the bitmask.

Fix the test, and remove the error message that would be printed when we
fail to parse a prefix length (parse_uint() already prints an error).

Remove curly braces, now useless.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>