parse: fix prefix length mask parsing for IPv6 addresses
The functions used to parse masks as prefix lengths for IPv6 addresses
are the same as for IPv4 addresses. However, they would consider that
the prefix length cannot be higher than 32, which is obviously incorrect
for IPv6.

Fix the functions accordingly. Also rename parse_slash_mask() as
parse_slash_prefix_mask() to avoid confusion with bitmask parsing, and
move it closer to parse_ip_addr_slash_mask(), which is its only caller
now.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>