parse_tc: clean up comments
File had some C++-style comment, used to indicate items to fix or
change. But those items have been fixed already, or will not be
addressed before a while, therefore we should update the comments to
better reflect the current state of the implementation, and to
consistently use C-style format.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>