parse_tc: enforce the use of "ip_proto" for L4 "src_port" or "dst_port"
This is not really a constraint for our library, but more something we
impose to be more consistent with the implementation of TC: flower rules
with L4 source or destination port patterns should also specify the L4
protocol to match in the IP header. Let's enforce this for our library
too, to remain as close as possible as what TC does.

Introduce a tcflower_check_matchlist() function that could be used to
perform additional verifications on the match objects list in the
future.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>