parse: fix prefix bitmask parsing for MAC addresses
While reusing the function for parsing unsigned integers seemed like a
good idea at the time, it stores the int depending on its maximal value,
and in our case, on one byte. However, bitmask_from_int() is expecting
an unsigned int. This is not working...

Do not reuse the parse_uint() function, call strtoul() again to parse
the bitmask from user input.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>