dump: fix update of appended text length
We should definitely subtract from, and not add to, the length of
remaining available space in the buffer when we append text to it.

This does not change the fact that if the buffer is too small, it does
not get reallocated for now, we just stop appending text to it.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>