proggen: manually unroll loop for checks if too many match types
When a filter uses a total number of match types (e.g. source IPv4
address, L4 protocol type, etc.) equal or greater than a certain number
(*), clang refuses to unroll the loops for checking the collected values
to see if they match the ones from the rule. This results in the BPF
program containing back edges and being rejected by the verifier.

Since we generate the C code anyway, we can just manually unroll the
loop when writing the relevant program. Drawbacks: the code from the
library, as well as the generated C program, are less easy to read.
Benefit: it works even with many match objects.

(*) That number seemed to be 10 at some points, although additional
tests showed that occasional failures to unroll would sometime happen
with 3 or more total different match types. Maybe some changes to other
parts of the library had an impact here? We want things to work, so
unroll at 3 or more.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>