proggen, dump: improve filter dump and use it in proggen
Improve to some extent the function used to dump a filter in text format
into a buffer, and use it to print (as comments) the filter that was
used to generate a BPF program at the end of the C code.

Reuse the buf_append() function created for proggen in the code supposed
to dump the filter. Change the API somewhat: the function supposed to
dump things to buffers now allocate buffers themselves, thus limiting
the risk of a user passing a buffer that would be too small.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>