examples: create examples directory and first example application
Create a new examples/ directory, to add some samples applications using
the library, and create one sample in there.

Additional example applications for the library should be added to this
directory in the future.

The new sample application is able to take some options passed by the
user to the command line. Note that all options are not used as of this
commit, some are expected to be used for tests that will be added in a
future commit (and should reuse the functions for parsing the options).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>