libkefir: make the API work with bpf_obj objects (libbpf)
Make the BPF API able to work with struct bpf_obj elements, in
particular offer function to destroy such BPF objects, to retrieve the
file descriptor of a program associated to such object, or to use it to
retrieve information about maps associated to that object. The latter,
in particular, seems extremely difficult to perform without a BPF
object, so we do not really have a choice but to accept one from users
if they want to load programs and fill maps in several steps.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>