libkefir, proggen: pass an attr object to kefir_convert_filter_to_cprog
The API and internal functions used to convert a filter object into a C
program take only two arguments: filter and target (XDP/TC). But there
may be other options that would be worth passing at that time. To keep
something that we can easily expand in the future, create a specific
struct kefir_cprog_attr to hold the different parameters to pass for C
code generation. Right now, it only contains the target, but more will
be added in the future.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>