libkefir, proggen: free buffer used for generating C program
When generating a C program fails, the functions would return without
freeing the newly-allocated buffer supposed to hold the C code, thus
creating a memory leak.

Free the buffer on error. Also free it on success, once it is no longer
necessary (e.g. once C code has been written to a file). Obviously, do
not free it when we are supposed to pass this buffer to the user.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>