libkefir: pass attr objects as const, take attrs for compiling to BPF
The function responsible for compiling the program from C to BPF (object
file) would take several parameters (for file names and clang/llc
binaries to use). This is not really extensible if we need to pass any
additional parameter in the future. Make it a dedicated attribute
object, and update wherever necessary.

Make this function, and the other ones using similar attribute objects
as arguments, take constant arguments (for those attributes). Update the
loading function so that it does not modify the attr directly (but
considers the ifindex NULL anyway when loading a non-offloaded program).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>