libkefir: add shortcut functions to convert, load, attach a filter
As an effort to provide simple, all-in-one functions that users can call
to turn a filter into a running XDP program with minimal effort, add two
API functions able to perform all the steps sequentially, without the
need for the user to use intermediary calls.

One of them only takes a filter and a ifindex, the second one allows
more flexibility by offering to pass the different attr objects that can
be used at each stage of the process.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>