internals: remove value format from data, make data the new value
The format of a value object is directly inferred from the match type of
the parent match object for that value. As we always have access to this
parent when processing the values, storing the format inside the value
is redundant information. We can delete it.

Change:

    struct kefir_value {
        union {} data;
        format;
    };

into:

    union kefir_value;

Since the "data" is the only member left in the struct kefir_value, we
can change that struct into the union only. It leads to a number of
changes throughout the library, but the resulting code gets simpler.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>