lib: remove match on ranges (not fully unsupported)
Although some elements for supporting matches on ranges of values, the
code generation is not implemented, and nothing is done with the
"max_value" field in match objects. Remove the elements related to that
feature, we can add it back later if support is required.

Note that matching on ranges can actually be achieved through comparison
operators (>=/<=) and several patterns to match, for example
L4 dst port >= 2000 && L4 dst port <= 8000.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>