libkefir: offer new API interface to build rules
In addition to the ethtool and flower tc rule parsers we had for
creating and building rules, expose to users a new way to create them.

This commit exposes the (previously) internal representation for the
rules to the user, so that they can build their own match and rule
objects, and than add them to a filter (filter structure is not
exposed).

Also create two helpers for helping with filling the match and rule
objects, in particular because the library requires the values and masks
in match objects to be stored in network byte order.

The helper creating match objects deduces the expected length for the
value it takes from the provided match type: also expose a function to
get length of value for a given match type, so that users can make sure
they pass a pointer to a long enough buffer.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>