json_save, json_restore: use stdout/stdin if filename is "-"
Make the functions saving/restoring filters to/from a file as JSON able
to write to, or to read from, standard output/input (all respectively)
if the file name provided is a simple dash.

Also mark kefir_filter_dump() as unstable in the header doc, and mention
kefir_filter_save_to_file() as a better alternative for producing
parse-able output.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>