tests: reuse options parsing from examples/, add HW offload tests
Reuse (soft link the files under tests/) the functions defined in the
examples for parsing options from the command line, in order to provide
a range of options for tests. Implement some options, which also allows
for some clean up in the code (fewer hard-coded parameters).

When requested through some of those options, try to offload the test
programs to the hardware. Because test runs are not supported by any
hardware at this date, we cannot run the programs on fake packets.
Therefore, we consider that the test succeeds if the program is
successfully loaded on the designated NIC, without actually trying to
run it.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>