examples: add tcflower2json and json2c programs
Add two programs to showcase JSON saving and restoring for the filters.

The first program, tcflower2json, parses its arguments as a TC flower
expression, produces a filter, and dumps it to the standard output as a
JSON filter.

The second program, json2c, loads a JSON filter object from a file,
turns it into a BPF-compatible C program, and dumps this C code into the
console.

If the file name passed to the second argument is "-", then the
application reads from standard input, and be piped onto the output from
tcflower2json.

This is not really the recommended way of using JSON filters, which do
not have documented specifications for being reused by the users and
should rather be limited to save/reload, but it's a fun example anyway.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>