examples, tests: add README files documenting examples and tests
Add documentation in the examples/ and tests/ repository, explaining how
the sample applications and the tests are suposed to be used.

No documentation for tests and examples will be provided in doc/, which
will instead point to the README.md files added in this commit. This is
because those items are not part of the library per se, they are simply
provided for testing or discovering the features offered by libkefir.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>