parse_tc, proggen: add support for CVLAN, SVLAN, make TC use SVLAN
TC flower accepts "vlan_*" and "cvlan_*" keywords.

Ethtool ntuples accepts "vlan*" keywords too.

But, lo! Not all VLANs are equal. As far as I understand, TC flower
uses "vlan_*" for the first VLAN tag (the actual sVLAN if QinQ is used),
while ethtool understands it as "any VLAN tag in packet".

This commit adds support for CVLAN_* and SVLAN_* match types. VLAN_* is
still present and means either CVLAN or SVLAN. Make TC flower "vlan_*"
use the "SVLAN_*" types.

The generated code is not completely optimized for CVLAN and SVLAN
types: we always try to get fields for up to two VLAN headers, here for
the sake of readability.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>