tests: use command line option to run only subset of tests
When the -t|--test_list option is provided, run only the tests provided
by the user in the comma-separated list on the command line. This allows
for running just a few tests instead of the whole suite. This is
particularly useful for debugging a broken test, and even better to
debug it without creating too much noise in the tracing logs (when using
printk()).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>