compile: add an API function to dump commands for updating maps
In case users want to update map themselves, for instance because they
modified the C program before loading it and want to update rules too,
provide a function to dump information about how the BPF map for rules
should be filled.

Dump the information in "bpftool map update" commands format.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>