libkefir, proggen: account for BPF bounded loops support
As of this commit, BPF support for bounded loops just landed in the
kernel. This means that we do not have to force unrolling of all loops
at compile time, instead we can have real (bounded) loops in the
programs!

As support for loops will be the default on Linux soon, let's consider
that bounded loops are supported by default. But since users may not use
the latest kernels just yet, add a cprog option to tell the library not
to account for bounded loops support (and to keep using the clang unroll
pragma instead).

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>