libkefir.h, internals: use _BITUL() macro i.o. defining/using bit()
The bit() macro was defined and used on a suggestion from checkpatch, as
it is the recommanded way to define flag values in the kernel. Since it
is not exposed to user space, we had to redefine it.

Instead, we could use something very similar from linux/const.h, which
is available in user space already: _BITUL(). Let's switch to that macro
instead of redefining things.

Suggested by @kuba-moo.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>