proggen: add an API option to change the license for C program
The C program has a license string placed in a specific ELF section,
that is passed to the kernel at load time and may be used to determine
if the BPF program is authorized to use the helpers it may call.

This license string defaults to "Dual BSD/GPL", because it seems
relevant for a program generated by a library under similar licenses.
But some usres may want to use something else, so provide an option to
change it when creating the C program.

Suggested by @kuba-moo.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>