libkefir.h: move struct kefir_load_attr just above function using it
The struct kefir_load_attr was initially defined just on top of the
first function that uses it in the header file, just as for the other
similar struct kefir_*_attr.

However, at some point the kefir_bpfobj_* function were added to the
header, right between that struct and the function using it. This was a
mistake: move those two functions above the struct kefir_load_attr, so
that the latter remain close to kefir_cprog_load_to_kernel().

Reported by @pcasconnetronome.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>